[hotfix][build] Move antrun jar unpacking to compile phase
The execution order of plugins running in the same phase is determined by the order in which they are defined in the appear.
This disastrous design decision is easy to handle in projects with a simple single-module (because the pom is what it is), but in complex multi-module projects (with inheritance, plugin-management sections, profiles, ...) it becomes ever increasingly difficult to make a sane judgement about the order in which plugins are defined, and thus executed.

This commit ensures that the antrun-plugin is executed after the dependency-plugin by moving the execution into the next phase (compile).
Technically speaking this is the wrong phase (since we're not compiling anything), but it is the lesser evil.